Now passing single filter tests, still need to work a bit on the chains
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/branches/filters_dynamic_loading@22574 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 92afaffd792deca520f5cd47958a8189358ce593

+++ stacks/common/filters/include/filters/mean.h
@@ -198,7 +198,7 @@ bool MultiChannelMeanFilter<T>::update(const std::vector<T> & data_in, std::vect
   //ROS_ASSERT(data_out.size() == width_);
   if (data_in.size() != number_of_channels_ || data_out.size() != number_of_channels_)
   {
-    ROS_ERROR("Configured with wrong size config:%d in:%d out:%d", number_of_channels_, data_in.size(), data_out.size());
+    ROS_ERROR("Configured with wrong size config:%d in:%d out:%d", number_of_channels_, (int)data_in.size(), (int)data_out.size());
     return false;
   }
